requirements.txt
psycopg2==2.9.9
cython
numpy
pandas
pyserial
streamlit
